Move audio device settings to modem settings area.
